Extract libjxl into its own Android Library
