+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "app/src/main/cpp/third_party/libjxl"]
-	path = app/src/main/cpp/third_party/libjxl
+	path = libjxl/src/main/cpp/third_party/libjxl
 	url = https://github.com/libjxl/libjxl.git
