+++ libjxl/src/main/java/fr/oupson/libjxl/JxlDecoder.java
@@ -1,11 +1,11 @@
-package fr.oupson.jxlviewer;
+package fr.oupson.libjxl;
 
 import android.graphics.drawable.AnimationDrawable;
 
 public class JxlDecoder {
+    public static native AnimationDrawable loadJxl( byte[] data);
+
     static {
         System.loadLibrary("jxlreader");
     }
-
-    public native AnimationDrawable loadJxl(byte[] data);
 }
