+++ settings.gradle
@@ -15,3 +15,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "JxlViewer"
 include ':app'
+include ':libjxl'
